Rationalize test file structure
